split logic and rendering for TaskList component
